+++ resolwe_bio/processes/import_data/metadata.py
@@ -87,12 +87,14 @@ class UploadOrangeMetadata(Process):
     slug = "upload-orange-metadata"
     name = "Metadata table for Orange"
     process_type = "data:metadata:orange"
-    version = "1.1.0"
+    version = "1.1.1"
     category = "Import"
     scheduling_class = SchedulingClass.BATCH
     requirements = {
         "expression-engine": "jinja",
-        "executor": {"docker": {"image": "resolwebio/orange:2.0.0"}},
+        "executor": {
+            "docker": {"image": "public.ecr.aws/s4q6j6e8/resolwebio/orange:2.0.0"}
+        },
         "resources": {"cores": 1, "memory": 8192},
     }
     data_name = '{{ src.file|default("?") }}'