+++ resolwe_bio/processes/import_data/sra_file.py
@@ -52,13 +52,15 @@ class ImportSra(Process):
     slug = "import-sra"
     name = "SRA data"
     process_type = "data:sra"
-    version = "1.4.1"
+    version = "1.4.2"
     category = "Import"
     scheduling_class = SchedulingClass.BATCH
     persistence = Persistence.TEMP
     requirements = {
         "expression-engine": "jinja",
-        "executor": {"docker": {"image": "resolwebio/sra-tools:2.0.0"}},
+        "executor": {
+            "docker": {"image": "public.ecr.aws/s4q6j6e8/resolwebio/sra-tools:2.0.0"}
+        },
         "resources": {
             "cores": 1,
             "memory": 8192,
@@ -147,7 +149,7 @@ class ImportSraSingle(Process):
     slug = "import-sra-single"
     name = "SRA data (single-end)"
     process_type = "data:reads:fastq:single"
-    version = "1.4.1"
+    version = "1.4.2"
     category = "Import"
     scheduling_class = SchedulingClass.BATCH
     persistence = Persistence.RAW
@@ -157,7 +159,9 @@ class ImportSraSingle(Process):
     }
     requirements = {
         "expression-engine": "jinja",
-        "executor": {"docker": {"image": "resolwebio/sra-tools:2.0.0"}},
+        "executor": {
+            "docker": {"image": "public.ecr.aws/s4q6j6e8/resolwebio/sra-tools:2.0.0"}
+        },
         "resources": {
             "cores": 1,
             "memory": 8192,
@@ -313,7 +317,7 @@ class ImportSraPaired(Process):
     slug = "import-sra-paired"
     name = "SRA data (paired-end)"
     process_type = "data:reads:fastq:paired"
-    version = "1.4.1"
+    version = "1.4.2"
     category = "Import"
     scheduling_class = SchedulingClass.BATCH
     persistence = Persistence.RAW
@@ -323,7 +327,9 @@ class ImportSraPaired(Process):
     }
     requirements = {
         "expression-engine": "jinja",
-        "executor": {"docker": {"image": "resolwebio/sra-tools:2.0.0"}},
+        "executor": {
+            "docker": {"image": "public.ecr.aws/s4q6j6e8/resolwebio/sra-tools:2.0.0"}
+        },
         "resources": {
             "cores": 1,
             "memory": 8192,