+++ resolwe_bio/tests/processes/test_chipseq.py
@@ -593,7 +593,7 @@ class ChipSeqProcessorTestCase(BioProcessTestCase):
                     },
                     "executor": {
                         "docker": {
-                            "image": "public.ecr.aws/s4q6j6e8/resolwebio/base:ubuntu-20.04",
+                            "image": "public.ecr.aws/s4q6j6e8/resolwebio/base:ubuntu-20.04-03042021",
                         },
                     },
                 },
@@ -647,7 +647,7 @@ re-save build "hg19"
                     },
                     "executor": {
                         "docker": {
-                            "image": "public.ecr.aws/s4q6j6e8/resolwebio/base:ubuntu-20.04",
+                            "image": "public.ecr.aws/s4q6j6e8/resolwebio/base:ubuntu-20.04-03042021",
                         },
                     },
                 },